Remove intermediary AuthResourceStore class
